fixing datetime and bit issue in DW. (#1931)
## Why make this change?

Datawarehouse was having issues with datetime. This is because we are
using string agg. if we dont add the "" to datetime, it is being
deserialized as a number and we get a bug saying there is a - in the
number. FOR JSON converts bit to true or false and hence we need to do
same.

## What is this change?

Hence for datetime we need to add the wrapping "" to ensure correct
deserialization.
if datatype is bit set to true or false.


## How was this tested?

unit tests added. 
integration test done with dab running and dw datetime supported. image:
![Uploading image.png…]()

---------

Co-authored-by: aaronburtle <93220300+aaronburtle@users.noreply.github.com>