Add support for app insights in CLI (#1952)
## Why make this change?

- Closes #1745 
- This change will allow users to add application insights to config
through CLI

## What is this change?

- Added a new cli command `add-telemetry` to add telemetry resource
where the data will be sent. Currently DAB supports sending telemetry to
Application Insights.
- this command has two flags/options:
`--app-insights-enabled`: A boolean flag that specifies whether
Application Insights telemetry should be enabled. This flag is optional
and default value is true.
`--app-insights-conn-string`: A string that specifies the connection
string for the Application Insights resource to which telemetry data
should be sent. This flag is required must be set to a valid connection
string.

- This command will add application insights telemetry resource details
in the config, when dab engine starts with this config it will send
telemetry details to applications insights.

## How was this tested?
- [X] Unit Tests

---------

Co-authored-by: Aniruddh Munde <anmunde@microsoft.com>
Co-authored-by: Shyam Sundar J <shyamsundarj@microsoft.com>