Updated Json Schema which includes `cache` property (#2011)
## Why this change?

This change follows up on
https://github.com/Azure/data-api-builder/pull/1865, which adds caching
support.
VSCode suggesting the new `cache` properties were errors when included
in the config file. That is because the properties needed to be added to
the json schema.

Global `cache` property in the `runtime` section which adjacent to
`rest`, `graphql`, and `host`.

```json
{
  "runtime": {
    "cache": {
      "enabled": false,
      "ttl-seconds": 5
    }
  }
}
```

and per entity alongside `source`, `graphql`, `rest`, and `permissions`

```json
{
  "myEntity": {
    "cache": {
      "enabled": false,
      "ttl-seconds": 5
    }
  }
}
```