GraphQL stored procedure support for DW (#2008)
## Why make this change?
This change adds stored procedure execution support for Datawarehouse. 

## What is this change?
In this change, the definitions in DwSqlQueryBuilder are created to have
the dw queries to get result set information of the stored procedure and
to actually execute stored procedure queries.

In addition, for both mssql and dwsql the execution logic flows through
a single parser. It ensures that both mssql and dw sql use consistent
naming to ensure easy parsing of information.

## How was this tested?

- [ ] Integration Tests
1.Tested against a warehouse:

![image](https://github.com/Azure/data-api-builder/assets/124841904/fac592c5-d99e-4d12-ac31-7c3ee043c641)
2. Added Tests for stored procedure.