[CherryPick] Create a unittest pipeline, moving common steps into templates, disab… (#2070)
…le CosmosDB testing on linux (#2049)

## Why make this change?

- Trying to improve the pipeline triggering experience by moving all PR
validation pipelines to public project.
- Desired goal is to have all pipelines automatically trigger upon
commenting `/azp run` on an approved PR by a team member

## What is this change?

- Create a unittest pipeline to run unittests
- Move common steps and variables from original build pipeline to
templates
- Remove testing on the linux platform for CosmosDB in preparation to
run this pipeline in the public project too. Its been difficult to start
a CosmosDB emulator on linux, preventing us to test on this platform in
the public project where pipeline secrets aren't allowed for PRs from
forked repositories. (they are allowed in private projects, hence we had
been able to test on linux when the pipeline was in the private project)

## How was this tested?

- Temporarily renamed the unittest-pipelines.yml as build-pipelines.yml
and ran the `Build` pipeline on this branch to verify a successful build



![image](https://github.com/Azure/data-api-builder/assets/3513779/12c250a0-5cbc-4687-b918-984e8f6e63f0)

Co-authored-by: Aniruddh Munde <anmunde@microsoft.com>