Stored proc support test cases REST DW. (#2048)
## Why make this change?
This change adds appropriate test cases that validates the REST support
for stored procedures for Data warehouse.

## What is this change?
This change adds test cases needed to validate stored proc rest support
in dab. It only adds test cases currently supported for other db types.
1. Find: For find, we remove the overriden test cases in the dab class.
this means the code will execute the test cases in the base classes for
DW.
2. Delete: Add the delete class for dwtests that inherits off of the
base. (delete support exists for rest dw case as for delete we return no
content as opposed to insert). The specific test cases that need to be
overriden due to DW specific implementations have been overriden and
added to the delete test cases file.
3. PUT/PATCH: There are no stored proc specific tests for PUT/PATCH for
other db's and hence none added here. Tracked in this issue:
https://github.com/Azure/data-api-builder/issues/2085
4. Insert: There is one test case pertaining to stored proc insert that
is added for other db's. That is not included in this pr, as since we
would want to inherit from the InsertTestBase class, other insert tests
would also execute. As compared to delete, insert on rest, currently
expects the inserted object to be returned. In DW, we have gone with
option of only returning success/failure for mutation operations instead
of the result and that logic + test cases will be implemented in
https://github.com/Azure/data-api-builder/issues/1982

## How was this tested?
test cases added.

---------

Co-authored-by: Aniruddh Munde <anmunde@microsoft.com>