Add validation for source and target fields in relationships (#2139)
## Why make this change?

Closes https://github.com/Azure/data-api-builder/issues/1935

By adding validation for source and target fields and linking source and
linking target fields in the `RuntimeConfigValidator`.

## What is this change?

In the `ValidateRelationshipsInConfig` method within the
`RuntimeConfigValidator`, we add logic to cover the cases from
https://github.com/Azure/data-api-builder/issues/1935 such that we
properly validate the relationships in the config.



## How was this tested?

- [X] Unit Tests

We add to the `ConfigValidationUnitTests` class all of the cases from
the above issue. We assert that the correct kind of exception is thrown,
with the correct message, status code, and sub status code.

## Sample Request(s)

You can manually test this by creating configs that violate the rules
outlined in the above issue. Or by stepping through the new unit tests
which do the same, but programmatically in the tests.