Adding pagination limits. (#2153)
## Why make this change?

This change allows users to use the runtime config to provide limits on
the number of records that can be retrieved through paginated calls. It
also gives them the ability to define a default size which will be the
size returned when no pagination input is given (no first in case of
graphql and no limit in case of rest calls).

## What is this change?
It adds the nullable paginationoptions property to runtimeoptions of
runtimeConfig.
Default page size is set to 100. 
Max page size is set to 100,000 ( can be altered by the user in their
runtimeconfig).

A call with -1 pagination input will result in max page size records
being returned. Any call with a pagination number higher than max page
size will be rejected with bad request.

In config:
default-page-size is an integer
default-page-size default is 100
default-page-size value -1 means "same as max-page-size"
default-page-size value 0 is an error
default-page-size value less than -1 is an error
default-page-size value more than than max-page-size is an error
max-page-size is an integer
max-page-size default is 100,000
max-page-size value -1 means "same as int.MaxValue"
max-page-size value 0 is an error
max-page-size value less than -1 is an error
max-page-size value more than int.MaxValue is an error

In a query (REST or GQL):
$first=-1 means "whatever the max-page-size value is"
$first=less than -1 is an error
$first=0 is an error
$first=(any value more than max-page-size) is an error

Sample configuration file:
```json
{
  "runtime": {
    "pagination": {
      "default-page-size": -1,
      "max-page-size": 1233
    }
  }
}
```

## How was this tested?
1. For default page value of 100, invalid page value of 0 or <-1 the
existing test cases should cover.
2. Added tests for both GQL and REST for above conditions.

---------

Co-authored-by: Sean Leonard <sean.leonard@microsoft.com>
Co-authored-by: aaronburtle <93220300+aaronburtle@users.noreply.github.com>